policy: Support listener priority on datapath
Add support for listener priority by adding a datapath policy map
ProxyPortPriority field.

This makes it possible to use HTTP policies with port ranges.

Move MapStateEntry definition into policy/types, so that it's components
are easier to import without causing merge conflicts.

+++ pkg/policy/api/l4.go
@@ -160,7 +160,7 @@ type Listener struct {
 	// +kubebuilder:validation:Minimum=1
 	// +kubebuilder:validation:Maximum=100
 	// +kubebuilder:validation:Optional
-	Priority uint16 `json:"priority"`
+	Priority uint8 `json:"priority"`
 }
 
 // PortRule is a list of ports/protocol combinations with optional Layer 7
